feat: Load records in tab child.
